requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.32
pydantic==2.7.4
pydantic-settings==2.3.4
pypdf==4.2.0
beautifulsoup4==4.12.3
langchain-core==0.2.40
langgraph==0.2.32
pgvector==0.2.5
psycopg2-binary==2.9.9
transformers==4.45.2
